Add user id tokens to lang_set messages
Replace the usernames in the lang_set message with user tokens of the
format $userid. Now if a translator does not have a username it will
fill in their name instead and the message will always have the most
current name/username when sent.

Change the usernames to the new tokens in lang.json