Handle an error that occurs when saving a sticker as a png
There is currently an error in the Pillow library that prevents saving
stickers that don't have a transparent background as png files.
I think the changes to fix this have already been committed to Pillow,
but unfortunately we have to wait until they release the next version
containing the patch.